rags/docs-rag/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
openai>=1.12.0
python-dotenv>=1.0.0
